refactor(web): move the shared editor and attachment parts to components/common
