+++ apps/web/messages/ar/common.json
@@ -63,9 +63,6 @@
       "issue": {
         "new": "مهمة جديدة"
       },
-      "initiative": {
-        "new": "مبادرة جديدة"
-      },
       "view": {
         "kanban": "تخطيط اللوحة",
         "table": "تخطيط الجدول",
