+++ apps/web/messages/en/common.json
@@ -63,9 +63,6 @@
       "issue": {
         "new": "New issue"
       },
-      "initiative": {
-        "new": "New initiative"
-      },
       "view": {
         "kanban": "Board layout",
         "table": "Table layout",
