+++ apps/web/messages/fr/common.json
@@ -63,9 +63,6 @@
       "issue": {
         "new": "Nouveau ticket"
       },
-      "initiative": {
-        "new": "Nouvelle initiative"
-      },
       "view": {
         "kanban": "Affichage tableau",
         "table": "Affichage liste",
