+++ apps/web/messages/ru/common.json
@@ -63,9 +63,6 @@
       "issue": {
         "new": "Новая задача"
       },
-      "initiative": {
-        "new": "Новая инициатива"
-      },
       "view": {
         "kanban": "Вид «Доска»",
         "table": "Вид «Таблица»",
