+++ apps/web/messages/uk/common.json
@@ -63,9 +63,6 @@
       "issue": {
         "new": "Нова задача"
       },
-      "initiative": {
-        "new": "Нова ініціатива"
-      },
       "view": {
         "kanban": "Вигляд «Дошка»",
         "table": "Вигляд «Таблиця»",
